requirements.txt
asgiref==3.8.1
Brotli==1.1.0
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==5.0.2
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-heroku==0.3.1
django-import-export==3.3.7
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-nested-admin==4.0.2
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
gunicorn==21.2.0
inflection==0.5.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-monkey-business==1.0.0
pytz==2024.1
PyYAML==6.0.1
setuptools==69.2.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0